refactor: update comment & optimize utils
