+++ packages/react-drager/package.json
@@ -47,11 +47,8 @@
   "devDependencies": {
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "autoprefixer": "^10.0.0",
-    "postcss": "^8.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tailwindcss": "^3.0.0",
     "tsup": "^8.0.0",
     "typescript": "^5.0.0"
   }
