+++ packages/react-drager/src/utils/dom.ts
@@ -1,0 +1,6 @@
+/**
+ * get all have data-drager attribute elements
+ */
+export function getDragerElements(): HTMLElement[] {
+  return Array.from(document.querySelectorAll('[data-drager]'))
+}
